+++ sample/feature/tiles/src/main/kotlin/com/mihaicristiancondrea/android/apps/apptoolkit/feature/tiles/ui/models/ToolkitTileModels.kt
@@ -25,7 +25,7 @@ import com.mihaicristiancondrea.android.apps.apptoolkit.feature.tiles.data.model
 import kotlinx.collections.immutable.ImmutableList
 
 enum class ToolkitTileIcon {
-    Level, Compass, Coin, Dice, Counter, Sound, Breathing, Sos, Morse, FlashDimmer, Palette,
+    Level, Compass, Coin, Dice, Counter, Sound, Breathing, Sos, Morse, FlashDimmer, Palette, Timer,
 }
 
 @Immutable